// lib/sec-api/model/userDTO.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { RoleDTO } from './roleDTO';


export interface UserDTO { 
    email?: string;
    id?: number;
    name?: string;
    password?: string;
    roles?: Array<RoleDTO>;
    username?: string;
}